Move TestProjectBase to utils.
